define one class in a one file
