+++ grisha.pletin/SixHW/Header.rb
@@ -1,0 +1,16 @@
+
+require './HTMLparse'
+
+class Header < Document
+  def parse
+    p 'Headers'
+    @page.css('h2').each do |el|
+      puts el.to_str
+    end
+    p '*********************************************************'
+  end
+end
+
+
+@header = Header.new
+@header.parse
