+++ grisha.pletin/SixHW/Link.rb
@@ -1,0 +1,15 @@
+
+require './HTMLparse'
+
+class Link < Document
+  def parse
+    p 'Links'
+    @page.css('a', 'href').each do |url|
+      puts url.to_str
+    end
+    p '*********************************************************'
+  end
+end
+
+@link = Link.new
+@link.parse
